requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
alembic>=1.13
pymysql>=1.1
python-dotenv>=1.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]
email-validator>=2.0.0